feat(showcase/google-adk/declarative-json-render): rename from byoc-json-render and port frontend from langgraph-python
