--- showcase/integrations/google-adk/src/app/demos/byoc-json-render/suggestions.ts
@@ -1,28 +1,0 @@
-/**
- * Suggestion prompts for the byoc-json-render demo.
- *
- * These match Wave 4a's (hashbrown) prompts verbatim so the two demos
- * can be compared side-by-side in the dashboard.
- */
-
-export interface Suggestion {
-  /** Short label shown on the pill. */
-  label: string;
-  /** Full prompt text sent to the agent when the pill is clicked. */
-  prompt: string;
-}
-
-export const BYOC_JSON_RENDER_SUGGESTIONS: Suggestion[] = [
-  {
-    label: "Sales dashboard",
-    prompt: "Show me the sales dashboard with metrics and a revenue chart",
-  },
-  {
-    label: "Revenue by category",
-    prompt: "Break down revenue by category as a pie chart",
-  },
-  {
-    label: "Expense trend",
-    prompt: "Show me monthly expenses as a bar chart",
-  },
-];
