+++ showcase/integrations/google-adk/src/app/demos/declarative-json-render/catalog.ts
@@ -1,0 +1,38 @@
+import { defineCatalog } from "@json-render/core";
+import { schema } from "@json-render/react/schema";
+import { z } from "zod";
+
+const dataPoint = z.object({ label: z.string(), value: z.number() });
+
+export const catalog = defineCatalog(schema, {
+  components: {
+    MetricCard: {
+      props: z.object({
+        label: z.string(),
+        value: z.string(),
+        trend: z.string().nullable(),
+      }),
+      description:
+        "A labelled metric (single number) with an optional trend subtitle",
+    },
+    BarChart: {
+      props: z.object({
+        title: z.string(),
+        description: z.string().nullable(),
+        data: z.array(dataPoint),
+      }),
+      description:
+        "A vertical bar chart for comparing discrete values side by side",
+    },
+    PieChart: {
+      props: z.object({
+        title: z.string(),
+        description: z.string().nullable(),
+        data: z.array(dataPoint),
+      }),
+      description:
+        "A donut-style pie chart for breaking a total down into category slices",
+    },
+  },
+  actions: {},
+});
