+++ showcase/integrations/google-adk/src/app/demos/declarative-json-render/charts/chart-config.ts
@@ -1,9 +1,3 @@
-/**
- * CopilotKit brand chart palette — Plus Jakarta Sans / brand color system.
- * Ported from `showcase/starters/template/frontend/components/charts/chart-config.ts`
- * to keep the byoc-json-render demo self-contained (per spec decision: each demo
- * maintains its own catalog component copies).
- */
 export const CHART_COLORS = [
   "#BEC2FF", // lilac-400
   "#85ECCE", // mint-400
