+++ showcase/integrations/google-adk/src/app/demos/declarative-json-render/types.ts
@@ -1,0 +1,7 @@
+export interface JsonRenderSpec {
+  root: string;
+  elements: Record<
+    string,
+    { type: string; props: Record<string, unknown>; children?: string[] }
+  >;
+}
